Fix use-after-free bug of DataType value.
Everytime when we call DataType<T>::value(), a new string is constructed and assign to the static data_type. So if we call it twice, the first one has been released when we get the second one.

std::string data_type(value1, strlen(value2));

If the parameters are evaluated from left to right, value1 becomes invalid while we are using it to construct data_type.
This is the simplest fix, which passes a single c_str as parameter.
A better solution is to refactor this whole piece of code, to try to avoid constructing data_type everytime, or avoid c_str.